Adds vlamax estimator
Uses an exponential fit for the power values to determine the rate of performance decay. The idea is the rate of decay is proportional to the rate with which lactate is produced. High decay means high lactate production rate. This is used mapped to measured vlamax values for the final result.